Re-implement I/O for AsyncRCONClient
AsyncRCONClient no longer cancels the current task, preferring
to propagate any exception raised by the protocol.
